requirements.txt
PyYAML==5.3.1
slumber==0.7.1
requests==2.22.0